Fix CloudFront logs parsing where Content-Length is provided as "-"
The "sc-content-len" field in the CloudFront logs format provides the value of
the "Content-Length" response header, as sent by the server.

This header is not a required header and in cases where it is not provided,
CloudFront instead provides the default value of "-" in the log files.

That does however not parse with the CloudFront parser as it expects the field
to always have a number in this column.

This change fixes that by instead changing the column to be parsed via the
"DATA" grok pattern instead of "NUMBER".

An extra test line was added to check this case.